+++ android/expoview/src/main/java/versioned/host/exp/exponent/ExperiencePackagePicker.kt
@@ -17,7 +17,7 @@ import expo.modules.clipboard.ClipboardModule
 import expo.modules.constants.ConstantsPackage
 import expo.modules.contacts.ContactsPackage
 import expo.modules.core.interfaces.Package
-import expo.modules.crypto.CryptoPackage
+import expo.modules.crypto.CryptoModule
 import expo.modules.device.DevicePackage
 import expo.modules.documentpicker.DocumentPickerPackage
 import expo.modules.easclient.EASClientModule
@@ -81,7 +81,6 @@ object ExperiencePackagePicker : ModulesProvider {
     CameraPackage(),
     ConstantsPackage(),
     ContactsPackage(),
-    CryptoPackage(),
     DevicePackage(),
     DocumentPickerPackage(),
     ErrorRecoveryPackage(),
@@ -143,6 +142,7 @@ object ExperiencePackagePicker : ModulesProvider {
   override fun getModulesList(): List<Class<out Module>> = listOf(
     CellularModule::class.java,
     ClipboardModule::class.java,
+    CryptoModule::class.java,
     EASClientModule::class.java,
     LinearGradientModule::class.java,
     RandomModule::class.java,