+++ packages/expo-crypto/src/ExpoCrypto.web.ts
@@ -1,6 +1,8 @@
-import { CodedError } from 'expo-modules-core';
+import { CodedError, TypedArray } from 'expo-modules-core';
 
 import { CryptoDigestAlgorithm, CryptoEncoding, CryptoDigestOptions } from './Crypto.types';
+
+const getCrypto = (): Crypto => window.crypto ?? (window as any).msCrypto;
 
 export default {
   get name(): string {
@@ -29,13 +31,14 @@ export default {
   },
   getRandomBytes(length: number): Uint8Array {
     const array = new Uint8Array(length);
-    // @ts-ignore
-    return (window.crypto ?? window.msCrypto).getRandomValues(array);
+    return getCrypto().getRandomValues(array);
   },
   async getRandomBytesAsync(length: number): Promise<Uint8Array> {
     const array = new Uint8Array(length);
-    // @ts-ignore
-    return (window.crypto ?? window.msCrypto).getRandomValues(array);
+    return getCrypto().getRandomValues(array);
+  },
+  getRandomValues(typedArray: TypedArray) {
+    return getCrypto().getRandomValues(typedArray);
   },
 };