+++ packages/expo-crypto/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### 🎉 New features
 
+- The module on Android now uses JSI host object instead of the bridge module for communication between JavaScript and native code. ([#17614](https://github.com/expo/expo/pull/17614) by [@lukmccall](https://github.com/lukmccall))
+
 ### 🐛 Bug fixes
 
 ### 💡 Others